fixed editsvg click event + some refactoring
