+++ src/main/html/bookshelf.html
@@ -43,8 +43,6 @@
                                                 <option value="bg-orange">Orange</option>
                                                 <option value="bg-pink">Pink</option>
                                                 <option value="bg-cyan">Blue</option>
-                                                <option value="bg-bw-red">Red</option>
-                                                <option value="bg-green">Green</option>
                                                 <option value="bg-tan">Tan</option>
                                             </select>
                                         </div>
@@ -125,6 +123,9 @@
         <div class="bg-slate-100 rounded-lg pb-1 w-full">
             <a>Hello World</a>   
         </div>
+    </div>
+    <div id="new-shelf-container">
+
     </div>
 </body>
 </html>
