+++ src/main/ts/bookshelf/newshelf.ts
@@ -12,13 +12,15 @@ export const createNewShelf = (shelfName: string, colorClass: string): HTMLDivEl
         playlistName.textContent = `${shelfName}`;
 
     const playlistEdit: SVGSVGElement = document.createElementNS('http://www.w3.org/2000/svg', 'svg') as SVGSVGElement;
-        playlistEdit.classList.add('h-6', 'w-6', 'inline-block', 'fill-white', 'cursor-pointer');
+        playlistEdit.classList.add('edit-list', 'h-6', 'w-6', 'inline-block', 'fill-white', 'cursor-pointer');
         playlistEdit.setAttribute('xmlns', 'http://www.w3.org/2000/svg');
         playlistEdit.setAttribute('viewBox', '0 0 24 24');
-        playlistEdit.id = 'edit-list';
+    const playlistEditId = 'edit-list-' + new Date().getTime();
+        playlistEdit.id = playlistEditId;
     const pathEditElement: SVGPathElement = document.createElementNS('http://www.w3.org/2000/svg', 'path') as SVGPathElement;
         pathEditElement.setAttribute('d', 'M16.293 2.293a1 1 0 0 1 1.414 0l4 4a1 1 0 0 1 0 1.414l-13 13A1 1 0 0 1 8 21H4a1 1 0 0 1-1-1v-4a1 1 0 0 1 .293-.707l10-10 3-3zM14 7.414l-9 9V19h2.586l9-9L14 7.414zm4 1.172L19.586 7 17 4.414 15.414 6 18 8.586z')
         playlistEdit.appendChild(pathEditElement);
+        
 
     const playlistNavigation: HTMLSpanElement = document.createElement('span');
         playlistNavigation.classList.add('inline-block', 'mr-8');
